Remove any possible human accounts
For `@ninjawhale`'s bot army, most of the accounts post on Steemit once or twice and then go dead silent (except for voting with ninja).

There are a few accounts that are still active (e.g. a post/article in the past few weeks) so I deleted them as we don't want false positives or to upset any human users.